refactor: Separate fetching logic
